+++ src_front/components/messages/ChatUnbanRequest.vue
@@ -5,7 +5,7 @@
 		
 		<Icon name="unbanRequest" alt="notice" class="icon"/>
 
-		<div>
+		<div class="content">
 			<i18n-t scope="global" tag="span" :keypath="label">
 				<template #USER>
 					<a class="userlink" @click.stop="openUserCard(messageData.user, messageData.channel_id)">{{messageData.user.displayName}}</a>
@@ -15,7 +15,10 @@
 				</template>
 			</i18n-t>
 	
-			<div class="quote" v-if="messageData.message">{{ messageData.message }}</div>
+			<template v-if="messageData.message">
+				<div class="quote" @click.stop="censor=false">{{ messageData.message }}</div>
+				<div class="warning" v-if="messageData.isFlagByAutomod"><Icon name="alert" />Le message contient des termes retenus par l'automod</div>
+			</template>
 		</div>
 	</div>
 </template>
@@ -35,19 +38,23 @@ class ChatUnbanRequest extends AbstractChatMessage {
 	declare messageData:TwitchatDataTypes.MessageUnbanRequestData;
 
 	public label:string = "";
+	public censor:boolean = false;
 
 	public get classes():string[] {
 		let res:string[] = ["chatunbanrequest", "chatMessage", "highlight"]
 		if(this.messageData.accepted === true) res.push("success");
 		else res.push("error");
+		if(this.censor) res.push("censor");
 		return res;
 	}
 	
 	public mounted():void {
+		this.messageData.isFlagByAutomod = true;
 		if(this.messageData.isResolve) {
 			this.label = this.messageData.accepted? "chat.unban_request.accepted" : "chat.unban_request.refused";
 		}else{
 			this.label = "chat.unban_request.request";
+			this.censor = true;
 		}
 	}
 }
@@ -56,6 +63,30 @@ export default toNative(ChatUnbanRequest);
 
 <style scoped lang="less">
 .chatunbanrequest{
-	
+	.quote {
+		cursor: pointer;
+		transition: filter .25s;
+	}
+
+	&.censor {
+		.quote {
+			filter: blur(6px);
+			&:hover {
+				filter: blur(3px);
+			}
+		}
+	}
+
+	.warning {
+		color: var(--color-light);
+		font-weight: normal;
+		padding: .25em .5em;
+		background-color: var(--color-alert-fade);
+		border-radius: var(--border-radius);
+		.icon {
+			height: 1em;
+			margin-right: .5em;
+		}
+	}
 }
 </style>