+++ src_front/utils/twitch/EventSub.ts
@@ -1101,11 +1101,13 @@ export default class EventSub {
 			type:TwitchatDataTypes.TwitchatMessageType.UNBAN_REQUEST,
 			user:await StoreProxy.users.getUserFrom("twitch", event.broadcaster_user_id, event.user_id, event.user_login, event.user_name),
 			isResolve:false,
+			isFlagByAutomod:false,
 			message:"",
 		}
 		if(topic == TwitchEventSubDataTypes.SubscriptionTypes.UNBAN_REQUEST_NEW) {
 			event = event as TwitchEventSubDataTypes.UnbanRequestEvent;
 			message.message = event.text;
+			message.isFlagByAutomod = await TwitchUtils.checkAutomodFlag(message.message);
 
 		}else if(topic == TwitchEventSubDataTypes.SubscriptionTypes.UNBAN_REQUEST_RESOLVED) {
 			event = event as TwitchEventSubDataTypes.UnbanRequestResolveEvent;
@@ -1119,6 +1121,7 @@ export default class EventSub {
 			message.message		= event.resolution_text;
 			message.accepted	= event.status != "denied";
 		}
+
 		StoreProxy.chat.addMessage(message);
 	}